#590 moved PropertiesOf and its test to package scalar
