+++ src/main/java/org/cactoos/scalar/PropertiesOf.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.cactoos.iterable;
+package org.cactoos.scalar;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -31,9 +31,10 @@ import org.cactoos.Input;
 import org.cactoos.Scalar;
 import org.cactoos.Text;
 import org.cactoos.io.InputOf;
+import org.cactoos.iterable.IterableOf;
+import org.cactoos.iterable.Mapped;
 import org.cactoos.map.MapEntry;
 import org.cactoos.map.MapOf;
-import org.cactoos.scalar.IoCheckedScalar;
 import org.cactoos.text.TextOf;
 
 /**
