+++ src/test/java/org/cactoos/scalar/PropertiesOfTest.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.cactoos.iterable;
+package org.cactoos.scalar;
 
 import java.security.SecureRandom;
 import java.util.Properties;
